adds league stat module and test
